Upload dataset and Jupyter notebook for passenger forecasting at major Indonesian airports
This commit includes the dataset of passenger numbers for four major airports in Indonesia (Soekarno-Hatta, Ngurah Rai, Juanda, and Kualanamu) from 2006 to 2024, and the Jupyter notebook `ARIMA_Bandara_YUNUS.ipynb`. The notebook applies SARIMA model for predicting passenger traffic for 2025. The dataset is used for time series forecasting and is essential for analyzing trends and making future predictions.